fixed job error/silenced processPendingUpdate jobs

+++ config/telescope.php
@@ -78,7 +78,7 @@ return [
     'queue' => [
         'connection' => env('TELESCOPE_QUEUE_CONNECTION', null),
         'queue' => env('TELESCOPE_QUEUE', null),
-        'delay' => env('TELESCOPE_QUEUE_DELAY', 10),
+        'delay' => (int) env('TELESCOPE_QUEUE_DELAY', 10),
     ],
 
     /*
